Change that Rick to better Rick
The only Rick I respect right now is the Rick from Rick & Morty. To be fair, you have to have a very high IQ to understand Rick and Morty. The humor is extremely subtle, and without a solid grasp of theoretical physics most of the jokes will go over a typical viewer's head.